requirements.txt
Flask
mysql-connector-python
gunicorn

django>=5.2,<5.3
gunicorn>=23,<24
dj-database-url>=3,<4
whitenoise>=6,<7
requests